Simplify redux implementation by using Ducks pattern
Action types, action creators, and reducers of the same domain are placed in
the same file.
