+++ client/src/containers/Dog/DogTable.test.js
@@ -1,7 +1,7 @@
 import DogTable from "./DogTable";
 import { render, screen } from "@testing-library/react";
 import { Provider } from "react-redux";
-import store from "../../redux/store";
+import store from "../../store/store";
 
 describe("DogTable", () => {
   it("should render dog table with headers", async () => {
