--- client/src/redux/actionTypes.js
@@ -1,1 +1,0 @@
-export const GET_ALL_DOGS_SUCCESS = "GET_ALL_DOGS_SUCCESS";
